refactor : Pet Entity type -> Pet Value type
중요 정보가 아니므로 값 타입을 변경
