+++ src/main/java/com/project/pet/user/service/UserService.java
@@ -76,9 +76,6 @@ public class UserService  {
         checkActBlackList(authentication);
 
         TokenInfo tokenInfo = jwtTokenProvider.generateToken(authentication); // 토큰 정보 발급
-
-
-        //log.info("Authentication = {}", authentication);
         /**
          * redis에 refresh token 저장! como에서는 jpa data를 활용한 repository를 만들어 저장했지만 여기서는
          * redisTemplate이 제공하는 set 메소드를 통해 직접 저장. 각 방법에 장단점이 있음. 여기서는 빠른 사용을 위해 해당 방법 채택
@@ -91,12 +88,7 @@ public class UserService  {
 
     private void checkActBlackList(Authentication authentication) {
         if(redisTemplete.hasKey("AT:" + authentication.getName())){
-            log.info("블랙리스트에 존재");
             redisTemplete.delete("AT:" + authentication.getName());
-            log.info("블랙리스트에 있는 \"AT:사용자아이디\" 삭제완료");
-        }
-        else {
-            log.info("해당 아이디에 대한 블랙리스트는 존재하지 않는다.");
         }
     }
 
@@ -121,53 +113,41 @@ public class UserService  {
      * 1. rft이 존재한다면 act을 재발급
      * 2. rtf이 존재하지 않는다면 로그인 필요
      *  예외처리 -> rtf 존재x, 로그인 필요
-     *
-     *
      */
     public String reissue(ReissueToken token) {
         String result = null;
-        //log.info("token 값 확인 = {}", token.getRefreshToken());
         String originRefreshToken = token.getRefreshToken();
         String resolvedToken = token.getRefreshToken().substring(7);
-        /*if (StringUtils.hasText(token) && token.startsWith("Bearer")) {
-            resolvedToken = token.substring(7);
-        }*/
-        System.out.println(resolvedToken);
-        //log.info("resolvedToken 값 재확인 = {}", resolvedToken);
+
         Claims claims = jwtTokenProvider.parseClaims(resolvedToken);
         String loginId = (String)claims.get("sub");
-        //System.out.println("loginId : " + loginId);
         String redisRft = (String) redisTemplete.opsForValue().get("RT:" + loginId); // 존재하지 않는다면 null 반환
 
-        //System.out.println("redisRft = " +redisRft);
-        //System.out.println("originRefreshToken = " + originRefreshToken);
-        //System.out.println(redisRft.equals(originRefreshToken));
         //클린 코드인가를 묻냐면 아닌 거 같다. 내 수준이 매우 낮으니 버그 발생이 없는 것을 목표로 빠르게 작성하고 리팩토링은 추후에 하자.
-
         // 아직 남아있는 redisRft가 있다면 redisRft와  rft를 비교, 사실 보안적으로 큰 향상은 없어 보이나 안 한 것보다 나으므로 적용
         if( redisTemplete.opsForValue().get("RT:" + loginId) != null && redisRft.equals(originRefreshToken)){
             long now = new Date().getTime();
             String reissueAct = jwtTokenProvider.reissueAccessToken(loginId, now);
-            /**
-             *
-             * 사실 이 재발급 과정은 매우 좋지 않다. jwtProvider이 제공하는 createAccessToken()이 있는데도 파라미터 변수가 매칭이 되지 않아
-             * 같은 기능의 코드를 파라미터만 바꿔 다시 만들었다. 마치 오버라이딩과 같이 사용했지만 중복된 코드이므로 가독성을 떨어뜨린다.
-             * 또한 형식도 Authentication을 통해 사용자 아이디와 역할(Role)을 얻는 것이 아니라 직접 아이디 값을 넣어주고
-             * reissueAccessToken() 내부에서 "일반 사용자"라고 고정하여 재생성한다.
-             * 만약 "일반 사용자"가 아닌 "관리자"인 경우에 의도치 않은 버그가 발생한다.
-             *
-             * 그런데도 이렇게 작성한 이유는? 완벽하게 하나씩 지식을 나의 것으로 확장시키는 것이 맞다고 생각하지만 조바심 떄문인지
-             * 빠르게 프로젝트를 완성시키고 리팩토링하려고 한다. 구글링으로는 원하는 지식을 빠르게 습득하기 어렵다고 판단하여(구글링 기술부족같음)
-             * 추후 강의 또는 세션을 통해 빠르게 리팩토링하겠다.
-             */
+
             result = reissueAct;
         }
         if(result == null){
             throw new IllegalArgumentException("저장된 rtf이 없거나 일치하지 않습니다.");
         }
-
         return result;
     }
+    /** 재발급 서비스 로직 문제점
+     *
+     * 사실 이 재발급 과정은 매우 좋지 않다. jwtProvider이 제공하는 createAccessToken()이 있는데도 파라미터 변수가 매칭이 되지 않아
+     * 같은 기능의 코드를 파라미터만 바꿔 다시 만들었다. 마치 오버라이딩과 같이 사용했지만 중복된 코드이므로 가독성을 떨어뜨린다.
+     * 또한 형식도 Authentication을 통해 사용자 아이디와 역할(Role)을 얻는 것이 아니라 직접 아이디 값을 넣어주고
+     * reissueAccessToken() 내부에서 "일반 사용자"라고 고정하여 재생성한다.
+     * 만약 "일반 사용자"가 아닌 "관리자"인 경우에 의도치 않은 버그가 발생한다.
+     *
+     * 그런데도 이렇게 작성한 이유는? 완벽하게 하나씩 지식을 나의 것으로 확장시키는 것이 맞다고 생각하지만 조바심 떄문인지
+     * 빠르게 프로젝트를 완성시키고 리팩토링하려고 한다. 구글링으로는 원하는 지식을 빠르게 습득하기 어렵다고 판단하여(구글링 기술부족같음)
+     * 추후 강의 또는 세션을 통해 빠르게 리팩토링하겠다.
+     */
 
     public Boolean checkDuplicateLoginId(String loginId) {
         return userRepository.findByLoginId(loginId).isPresent();
